--- app/src/main/kotlin/com/numero/sojodia/presenter/IUpdateBusDataPresenter.kt
@@ -1,7 +1,0 @@
-package com.numero.sojodia.presenter
-
-import com.numero.sojodia.core.Presenter
-
-interface IUpdateBusDataPresenter : Presenter {
-    fun checkUpdate()
-}
