+++ app/src/main/kotlin/com/numero/sojodia/service/UpdateBusDataPresenter.kt
@@ -1,0 +1,7 @@
+package com.numero.sojodia.service
+
+import com.numero.sojodia.core.Presenter
+
+interface UpdateBusDataPresenter : Presenter {
+    fun checkUpdate()
+}
