+++ app/src/main/kotlin/com/numero/sojodia/ui/board/BusSchedulePresenter.kt
@@ -1,229 +1,19 @@
 package com.numero.sojodia.ui.board
 
-import com.numero.sojodia.model.*
-import com.numero.sojodia.repository.BusDataRepository
+import com.numero.sojodia.core.Presenter
+import com.numero.sojodia.model.Route
+import com.numero.sojodia.model.Week
 
-class BusSchedulePresenter(
-        private val view: BusScheduleContract.View,
-        busDataRepository: BusDataRepository,
-        private val reciprocate: Reciprocate
-) : BusScheduleContract.Presenter {
-    private lateinit var week: Week
+interface BusSchedulePresenter:Presenter {
+    fun onTimeChanged(week: Week)
 
-    private var tkBusPosition = 0
-    private var tndBusPosition = 0
-    private val busData: BusData = busDataRepository.getBusData()
+    fun nextTkBus()
 
-    private var tkTimeList: MutableList<BusTime> = mutableListOf()
-        get() {
-            return if (field.isEmpty()) {
-                when (reciprocate) {
-                    Reciprocate.GOING -> busData.tkBusTimeListGoing
-                            .filter { busTime -> busTime.week == week }.toMutableList()
-                            .apply {
-                                field = this
-                            }
-                    Reciprocate.RETURN -> busData.tkBusTimeListReturn
-                            .filter { busTime -> busTime.week == week }.toMutableList()
-                            .apply {
-                                field = this
-                            }
-                }
-            } else {
-                field
-            }
-        }
+    fun previewTkBus()
 
-    private var tndTimeList: MutableList<BusTime> = mutableListOf()
-        get() {
-            return if (field.isEmpty()) {
-                when (reciprocate) {
-                    Reciprocate.GOING -> busData.tndBusTimeListGoing
-                            .filter { busTime -> busTime.week == week }.toMutableList()
-                            .apply {
-                                field = this
-                            }
-                    Reciprocate.RETURN -> busData.tndBusTimeListReturn
-                            .filter { busTime -> busTime.week == week }.toMutableList()
-                            .apply {
-                                field = this
-                            }
-                }
-            } else {
-                field
-            }
-        }
+    fun nextTndBus()
 
-    init {
-        this.view.setPresenter(this)
-    }
+    fun previewTndBus()
 
-    override fun subscribe() {}
-
-    override fun unSubscribe() {}
-
-    override fun onTimeChanged(week: Week) {
-        this.week = week
-        tkTimeList.clear()
-        tndTimeList.clear()
-        view.showTkBusTimeList(tkTimeList)
-        view.showTndBusTimeList(tndTimeList)
-
-        tkBusPosition = findBusPosition(tkTimeList)
-        tndBusPosition = findBusPosition(tndTimeList)
-
-        if (tkBusPosition == NO_BUS_POSITION) {
-            view.showTkNoBusLayout()
-            view.hideTkNextButton()
-            view.hideTkPreviewButton()
-        } else {
-            view.hideTkNoBusLayout()
-            view.selectTkCurrentBusPosition(tkBusPosition)
-            view.startTkCountDown(tkTimeList[tkBusPosition])
-            if (tkBusPosition >= tkTimeList.size - 1) {
-                view.hideTkNextButton()
-            } else {
-                view.showTkNextButton()
-            }
-            if (canPreviewTkTime()) {
-                view.showTkPreviewButton()
-            } else {
-                view.hideTkPreviewButton()
-            }
-        }
-        if (tndBusPosition == NO_BUS_POSITION) {
-            view.showTndNoBusLayout()
-            view.hideTndNextButton()
-            view.hideTndPreviewButton()
-        } else {
-            view.hideTndNoBusLayout()
-            view.selectTndCurrentBusPosition(tndBusPosition)
-            view.startTndCountDown(tndTimeList[tndBusPosition])
-            if (tndBusPosition >= tndTimeList.size - 1) {
-                view.hideTndNextButton()
-            } else {
-                view.showTndNextButton()
-            }
-            if (canPreviewTndTime()) {
-                view.showTndPreviewButton()
-            } else {
-                view.hideTndPreviewButton()
-            }
-        }
-    }
-
-    override fun nextTkBus() {
-        if (tkBusPosition == tkTimeList.size - 1) {
-            tkBusPosition = NO_BUS_POSITION
-        }
-        if (tkBusPosition == NO_BUS_POSITION) {
-            view.hideTkNextButton()
-            view.hideTkPreviewButton()
-            view.showTkNoBusLayout()
-            return
-        }
-        tkBusPosition += 1
-        if (tkBusPosition >= tkTimeList.size - 1) {
-            view.hideTkNextButton()
-        } else {
-            view.showTkNextButton()
-        }
-        if (canPreviewTkTime()) {
-            view.showTkPreviewButton()
-        } else {
-            view.hideTkPreviewButton()
-        }
-        view.selectTkCurrentBusPosition(tkBusPosition)
-        view.startTkCountDown(tkTimeList[tkBusPosition])
-    }
-
-    override fun previewTkBus() {
-        tkBusPosition -= 1
-        view.selectTkCurrentBusPosition(tkBusPosition)
-        view.startTkCountDown(tkTimeList[tkBusPosition])
-        if (canPreviewTkTime()) {
-            view.showTkPreviewButton()
-        } else {
-            view.hideTkPreviewButton()
-        }
-        if (tkBusPosition >= tkTimeList.size - 1) {
-            view.hideTkNextButton()
-        } else {
-            view.showTkNextButton()
-        }
-    }
-
-    override fun nextTndBus() {
-        if (tndBusPosition == tndTimeList.size - 1) {
-            tndBusPosition = NO_BUS_POSITION
-        }
-        if (tndBusPosition == NO_BUS_POSITION) {
-            view.hideTndNextButton()
-            view.hideTndPreviewButton()
-            view.showTndNoBusLayout()
-            return
-        }
-        tndBusPosition += 1
-        if (tndBusPosition >= tndTimeList.size - 1) {
-            view.hideTndNextButton()
-        } else {
-            view.showTndNextButton()
-        }
-        if (canPreviewTndTime()) {
-            view.showTndPreviewButton()
-        } else {
-            view.hideTndPreviewButton()
-        }
-        view.selectTndCurrentBusPosition(tndBusPosition)
-        view.startTndCountDown(tndTimeList[tndBusPosition])
-    }
-
-    override fun previewTndBus() {
-        tndBusPosition -= 1
-        view.selectTndCurrentBusPosition(tndBusPosition)
-        view.startTndCountDown(tndTimeList[tndBusPosition])
-        if (canPreviewTndTime()) {
-            view.showTndPreviewButton()
-        } else {
-            view.hideTndPreviewButton()
-        }
-        if (tndBusPosition >= tndTimeList.size - 1) {
-            view.hideTndNextButton()
-        } else {
-            view.showTndNextButton()
-        }
-    }
-
-    override fun showTimeTableDialog(route: Route) {
-        view.showTimeTableDialog(route, reciprocate)
-    }
-
-    private fun findBusPosition(busTimeList: List<BusTime>): Int {
-        busTimeList.forEachIndexed { index, busTime ->
-            if (busTime.time > Time()) {
-                return index
-            }
-        }
-        return NO_BUS_POSITION
-    }
-
-    private fun canPreviewTkTime(): Boolean {
-        if (tkBusPosition == NO_BUS_POSITION || tkBusPosition == 0) {
-            return false
-        }
-        val busTime = tkTimeList[tkBusPosition - 1]
-        return busTime.time > Time()
-    }
-
-    private fun canPreviewTndTime(): Boolean {
-        if (tndBusPosition == NO_BUS_POSITION || tndBusPosition == 0) {
-            return false
-        }
-        val busTime = tndTimeList[tndBusPosition - 1]
-        return busTime.time > Time()
-    }
-
-    companion object {
-        private const val NO_BUS_POSITION = -1
-    }
+    fun showTimeTableDialog(route: Route)
 }
