--- app/src/main/kotlin/com/numero/sojodia/ui/splash/ISplashPresenter.kt
@@ -1,5 +1,0 @@
-package com.numero.sojodia.ui.splash
-
-import com.numero.sojodia.core.Presenter
-
-interface ISplashPresenter : Presenter
