+++ app/src/main/kotlin/com/numero/sojodia/ui/splash/SplashView.kt
@@ -1,6 +1,6 @@
 package com.numero.sojodia.ui.splash
 
-interface ISplashView {
+interface SplashView {
 
     fun successDownloadedBusData()
